Very hacky fixes for shared server on local subnet.
I have a strange setup and PlexConnect wasn't working for me
as released.

I live in a building with a shared LAN. I do not own the Plex
Media Server, rather it has been shared with me.

I my network is 192.168.2.0/24, the plex server is on
192.168.1.0/24. GDM does not work for discovery.

Plex Media Server = 192.168.1.45

PlexConnect = 192.168.1.107
aTV = 192.168.1.102

What I discovered was PlexConnect was asking http(s)://plex.tv
for a list of servers. Plex.tv was handing back two different
IP address: the public IP address and the internal LAN addresses.
Next, PlexConnect would try and access the local Plex Media
Server by it's *public* address and not its local LAN address.
PlexConnect would be unable to connect to it. I fixed this by
hacking in the localAddresses returned from the plex.tv query.

Next, a number of the flags on the PMS server had to be over-ridden
to make PlexConnect think it was talking to a local owned server,
vs. a remote shared server.

Then, voila! Streaming shared entertainment!

In order to make these changes mergable back into the main
repository, they should probably be wrapped behind a configuration
option...